+++ practice/src/main/java/com/example/demo/config/SecurityConfig.java
@@ -4,14 +4,9 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.security.provisioning.InMemoryUserDetailsManager;
 import org.springframework.security.web.SecurityFilterChain;
-
 
 @EnableWebSecurity
 @Configuration
@@ -26,7 +21,7 @@ public class SecurityConfig {
             )
             .formLogin(form -> form
                 .loginPage("/admin/signin")
-                .defaultSuccessUrl("/admin/contacts", true) 
+                .defaultSuccessUrl("/admin/contacts", true)
                 .permitAll()
             )
             .logout(logout -> logout
@@ -40,15 +35,4 @@ public class SecurityConfig {
     public PasswordEncoder passwordEncoder() {
         return new BCryptPasswordEncoder();
     }
-
-    @Bean
-    public UserDetailsService userDetailsService() {
-        UserDetails user = User.withUsername("admin")
-            .password(passwordEncoder().encode("admin"))
-            .roles("ADMIN")
-            .build();
-
-        return new InMemoryUserDetailsManager(user);
-    }
 }
-
