--- practice/src/main/java/com/example/demo/controller/AdminLoginController.java
@@ -1,14 +1,0 @@
-package com.example.demo.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class AdminLoginController {
-
-    @GetMapping("/admin/signin")
-    public String showLoginForm() {
-        return "signin";
-    }
-
-}
