+++ practice/src/main/resources/application.properties
@@ -15,4 +15,7 @@ spring.jpa.properties.hibernate.format_sql=true
 spring.jpa.properties.hibernate.use_sql_comments=true
 logging.level.org.springframework.security=DEBUG
 logging.level.org.springframework.web=DEBUG
+spring.thymeleaf.prefix=classpath:/templates/
+spring.thymeleaf.suffix=.html
+
