+++ practice/src/main/resources/templates/signin.html
@@ -1,0 +1,20 @@
+<!DOCTYPE html>
+<html lang="ja">
+<head>
+    <meta charset="UTF-8">
+    <title>管理者ログイン</title>
+</head>
+<body>
+    <h1>管理者ログイン</h1>
+    <form action="/admin/signin" method="POST">
+        <label for="email">メールアドレス:</label>
+        <input type="email" name="email" required>
+        <br>
+        <label for="password">パスワード:</label>
+        <input type="password" name="password" required>
+        <br>
+        <button type="submit">ログイン</button>
+    </form>
+    <a href="/admin/signup">新規登録はこちら</a>
+</body>
+</html>
